pyproject.toml
[tool.poetry]
name = "rctcomms"
version = "0.0.13"
description = "Radio Collar Tracker Comms"
authors = [
    "UCSD Engineers for Exploration <e4e@ucsd.edu>",
    "Nathan Hui <nthui@ucsd.edu>"
]
readme = "README.md"
repository = 'https://github.com/UCSD-E4E/radio_collar_tracker_comms'

[tool.poetry.dependencies]
python = "^3.12"
pyserial = "^3.5"
schema = "^0.7.7"
deprecated = "^1.2.14"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.2"
autopep8 = "^2.3.1"
pylint = "^3.2.5"
isort = "^5.13.2"
pytest-timeout = "^2.3.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.semantic_release]
version_variables = ["RCTComms/__init__.py:__version__", "pyproject.toml:version"]
assets = []
commit_message = "{version}\n\nAutomatically generated by python-semantic-release"
commit_parser = "angular"
logging_use_named_masks = false
major_on_zero = true
tag_format = "v{version}"

[tool.semantic_release.branches.main]
match = "(main|master)"
prerelease = false
prerelease_token = "rc"

[tool.semantic_release.changelog]
template_dir = "templates"
changelog_file = "CHANGELOG.md"
exclude_commit_patterns = []

[tool.semantic_release.changelog.environment]
block_start_string = "{%"
block_end_string = "%}"
variable_start_string = "{{"
variable_end_string = "}}"
comment_start_string = "{#"
comment_end_string = "#}"
trim_blocks = false
lstrip_blocks = false
newline_sequence = "\n"
keep_trailing_newline = false
extensions = []
autoescape = true

[tool.semantic_release.commit_author]
env = "GIT_COMMIT_AUTHOR"
default = "semantic-release <semantic-release>"

[tool.semantic_release.commit_parser_options]
allowed_tags = ["build", "chore", "ci", "docs", "feat", "fix", "perf", "style", "refactor", "test"]
minor_tags = ["feat"]
patch_tags = ["fix", "perf"]

[tool.semantic_release.remote]
name = "origin"
type = "github"
ignore_token_for_push = false

[tool.semantic_release.remote.token]
env = "GH_TOKEN"

[tool.semantic_release.publish]
dist_glob_patterns = ["dist/*"]
upload_to_vcs_release = true
